Adicionando components, services e models no frontend e alterando tipo do atributo preço de BigDecimal para int no backend

+++ Desenvolvimento_Backend/projeto_final/loja-games-backend/src/main/java/com/programando/lojagamesbackend/model/Jogo.java
@@ -20,7 +20,7 @@ public class Jogo {
     private String descricao;
 
     @NotNull
-    private BigDecimal preco;
+    private int preco;
 
     @NotNull
     private String imagem;
